# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/apps/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /cell:
        post:
            tags:
                - Cell
            operationId: Cell_OnMessageReceived
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Reply'
    /chief:
        post:
            tags:
                - Chief
            operationId: Chief_OnMessageReceived
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Reply'
    /sys:
        post:
            tags:
                - System
            operationId: System_OnMessageReceived
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Reply'
components:
    schemas:
        Reply:
            type: object
            properties:
                requestId:
                    type: string
                code:
                    type: integer
                    format: uint32
                data:
                    type: string
                    format: bytes
                message:
                    type: string
        Request:
            type: object
            properties:
                requestId:
                    type: string
                operate:
                    type: integer
                    format: enum
                data:
                    type: string
                    format: bytes
tags:
    - name: Cell
    - name: Chief
    - name: System
      description: '****************************************** UserService *****************************************************'